Implement asynchronous queue and add some notes
The previous queue wasn't asynchronous, so `fs.write()` was actually
being called multiple times before the first one called the callback.

There was also an issue where we were using the queue without setting
`writing = true` on `get()`, but when I started looking into that issue
I realized that our queue wasn't working correctly in the first place
and decided to rewrite it to do what it was meant to do.

Whether *any* queue is necessary is up for debate, but I'm not
completely confident that Node.js handles all of this.